+++ src/ui/performance_renderer_groove.h
@@ -1,0 +1,11 @@
+/**
+ * @file performance_renderer_groove.h
+ * @brief Companion translation-unit boundary for PerformanceRenderer groove status drawing.
+ * @author Axel Napolitano
+ * @copyright 2026 Axel Napolitano
+ * @license PolyForm-Noncommercial-1.0.0
+ */
+
+#pragma once
+
+#include "ui/performance_renderer.h"