+++ src/ui/settings_editor_channel.h
@@ -1,0 +1,11 @@
+/**
+ * @file settings_editor_channel.h
+ * @brief Companion translation-unit boundary for flat Channel-menu mutation routing.
+ * @author Axel Napolitano
+ * @copyright 2026 Axel Napolitano
+ * @license PolyForm-Noncommercial-1.0.0
+ */
+
+#pragma once
+
+#include "ui/settings_editor.h"